file utils overhaul
