+++ src/utils/mod.rs
@@ -1,5 +1,4 @@
 mod files;
-mod impls;
 
 use anyhow::{anyhow, Context, Result};
 pub use files::*;
